pyproject.toml
[project]
name = "omni-sys"
version = "0.1.0"
description = ""
authors = [
    {name = "Jer Roy Tan",email = "ge48hac@mytum.de"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref (>=3.8.1,<4.0.0)",
    "django (>=5.2.1,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "pyjwt (>=2.9.0,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "hl7 (>=0.4.5,<0.5.0)",
    "pydicom (>=3.0.1,<4.0.0)",
    "hl7apy (>=1.3.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
